% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ci_handouts_pdf.R
\name{ci_handouts_pdf}
\alias{ci_handouts_pdf}
\title{Use the handout version of slides}
\usage{
ci_handouts_pdf(subtitle, location, institute, cover, cover_offset, fontsize,
  lang = "english", slide_level = 2, keep_tex = FALSE, ...)
}
\arguments{
\item{subtitle}{The date and place of the event}

\item{location}{The date and place of the event}

\item{institute}{The aviliation of the authors}

\item{cover}{The filename of the cover image}

\item{cover_offset}{An optional offset for the cover image}

\item{fontsize}{The fontsite of the document. Defaults to 10pt.}

\item{lang}{The language of the document. Defaults to "dutch"}

\item{slide_level}{Indicate which heading level is used for the frame titles}

\item{keep_tex}{Keep the tex file. Defaults to FALSE.}

\item{...}{extra parameters}
}
\description{
Use the handout version of slides
}

