% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ci_report_html.R
\name{ci_report_html}
\alias{ci_report_html}
\title{A function to convert an R Markdown document to a HTML page}
\usage{
ci_report_html()
}
\description{
A function to convert an R Markdown document to a HTML page
}

