% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/column_width.R
\name{column_start}
\alias{column_start}
\title{Calculate the starting position of a column}
\usage{
column_start(ncol = 3, start = 1, space = 20)
}
\arguments{
\item{ncol}{the required number of columns}

\item{start}{the number of the columns}

\item{space}{the space in between two columns in mm}
}
\description{
Calculate the starting position of a column
}
\examples{
column_start(ncol = 3, start = 2, space = 20)
}

