% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/column_width.R
\name{column_width}
\alias{column_width}
\title{Calculate the width of an A0 column}
\usage{
column_width(ncol = 3, space = 20)
}
\arguments{
\item{ncol}{the required number of columns}

\item{space}{the space in between two columns in mm}
}
\description{
Calculate the width of an A0 column
}
\examples{
column_width(ncol = 3, space = 20)
}

