% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ci_report_pdf.R
\name{ci_report_pdf}
\alias{ci_report_pdf}
\title{A function to convert an Rmarkdown document to a pdf report}
\usage{
ci_report_pdf(subtitle, cover, cover_offset, cover_text, reportnumber, lang,
  keep_tex = FALSE)
}
\description{
A function to convert an Rmarkdown document to a pdf report
}

